refactor: extract fetch logic to the user hook
